Upgrade to work with CKAN 2.9+
